data_alchemy_be/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
celery==5.3.6
django-celery-results==2.5.1
redis==5.0.1
pandas==2.1.4
python-dotenv==1.0.0
openpyxl==3.1.2